+++ examples/provider/provider-apikey.tf
@@ -1,0 +1,6 @@
+provider "elasticstack" {
+  elasticsearch {
+    api_key   = "base64encodedapikeyhere=="
+    endpoints = ["http://localhost:9200"]
+  }
+}