+++ internal/acctest/acctest.go
@@ -24,8 +24,14 @@ func PreCheck(t *testing.T) {
 	_, endpointsOk := os.LookupEnv("ELASTICSEARCH_ENDPOINTS")
 	_, userOk := os.LookupEnv("ELASTICSEARCH_USERNAME")
 	_, passOk := os.LookupEnv("ELASTICSEARCH_PASSWORD")
+	_, apikeyOk := os.LookupEnv("ELASTICSEARCH_API_KEY")
 
-	if !endpointsOk || !userOk || !passOk {
-		t.Fatal("ELASTICSEARCH_ENDPOINTS, ELASTICSEARCH_USERNAME, ELASTICSEARCH_PASSWORD must be set for acceptance tests to run")
+	if !endpointsOk {
+		t.Fatal("ELASTICSEARCH_ENDPOINTS must be set for acceptance tests to run")
+	}
+
+	usernamePasswordOk := userOk && passOk
+	if !((!usernamePasswordOk && apikeyOk) || (usernamePasswordOk && !apikeyOk)) {
+		t.Fatal("Either ELASTICSEARCH_USERNAME and ELASTICSEARCH_PASSWORD must be set, or ELASTICSEARCH_API_KEY must be set for acceptance tests to run")
 	}
 }